LibRequests+RequestServer: Store request IDs as u64
Not super important, but this will match an upcoming ID for stale-while-
revalidate requests. It also would techinically be UB if this ID had
ever overflowed. Let's make it 64-bit while we are here to also avoid
the possibility that it ever will overflow.

+++ Libraries/LibRequests/Request.cpp
@@ -24,7 +24,7 @@ ErrorOr<NonnullOwnPtr<ReadStream>> ReadStream::create(int reader_fd)
 #endif
 }
 
-Request::Request(RequestClient& client, i32 request_id)
+Request::Request(RequestClient& client, u64 request_id)
     : m_client(client)
     , m_request_id(request_id)
 {
